requirements.txt
kafka-python==2.0.2
load-dotenv==0.1.0
python-dotenv==1.0.0
